Rename overlay to image_overlay
