+++ tools/twix/src/panels/image/overlays/field_border.rs
@@ -6,7 +6,7 @@ use types::{field_border::FieldBorder as FieldBorderData, time_wrapper::TimeWrap
 
 use crate::repaint::ObservationContext;
 
-use super::super::overlay::{ImageOverlay, ImageOverlayPainter, OverlayObservation};
+use super::super::image_overlay::{ImageOverlay, ImageOverlayPainter, OverlayObservation};
 
 pub(in crate::panels::image) struct FieldBorderOverlay {
     border_lines: OverlayObservation<TimeWrapper<Option<FieldBorderData>>>,
