+++ tools/twix/src/panels/image/overlays/object_detection.rs
@@ -4,7 +4,7 @@ use types::object_detection::{Object, RobocupObjectLabel};
 
 use crate::repaint::ObservationContext;
 
-use super::super::overlay::{ImageOverlay, ImageOverlayPainter, OverlayObservation};
+use super::super::image_overlay::{ImageOverlay, ImageOverlayPainter, OverlayObservation};
 
 pub(in crate::panels::image) struct ObjectDetectionOverlay {
     object_detections: OverlayObservation<Vec<Object<RobocupObjectLabel>>>,
